+++ src/components/Utilities/sidebarNav.js
@@ -1,4 +1,4 @@
-import Profile from '@/components/Header/Profile/profile';
+import Profile from '@/components/Header/profile';
 import { XMarkIcon } from '@heroicons/react/24/solid';
 import Link from 'next/link';
